+++ src/compas_brep/__init__.py
@@ -1,6 +1,6 @@
 """compas_brep: Pure Python Brep implementation based on the COMPAS framework."""
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 
 import os